cc: give an enumeration a type wide enough for its enumerators
Every enum was four signed bytes, so an enumerator that did not fit was
truncated without a word:

    enum Big  { X = 5000000000 };   /* X == 705032704 */
    enum Huge { H = 0xFFFFFFFFU };  /* H == -1        */

C17 6.7.2.2p4 requires the enumerated type to represent every member.
`enum_underlying_type` now picks the narrowest integer type that does --
int, unsigned int, long, unsigned long, signed preferred at each width --
and the enumeration's size, alignment and its constants' type all follow
from it. An enum whose members fit in `int`, which is nearly all of them,
is unchanged.

Three places answered the width with a literal 4 (or 32 bits) rather than
asking the composite, so `sizeof(enum Big)` stayed 4 even once the type
was right. They ask now, defaulting to `int` for a forward reference that
has no members yet.

6.7.2.2p2 also makes an out-of-range enumerator a constraint violation,
which 5.1.1.3 requires be diagnosed. gcc widens without complaint unless
asked; c17 widens *and* warns, because silence is what turned a wrong
value into a running program. No system header on this machine trips it.

Enumerator values are held at 128 bits from parse onward -- the width
that `eval_const_expr` already returns -- so nothing wraps before the
range is known. An enumeration needing both ends of a range no integer
type spans is now an error instead of a coin flip.

+++ cc/ir/linearize_init.rs
@@ -348,7 +348,7 @@ impl<'a> super::linearize::Linearizer<'a> {
                     // Check if it's an enum constant
                     let sym = self.symbols.get(*symbol_id);
                     if let Some(val) = sym.enum_value {
-                        Initializer::Int(val as i128)
+                        Initializer::Int(val)
                     } else {
                         Initializer::None
                     }
